Print info with bitmapped font
Info strings get printed in the window using a bitmap font. Info level
can be cycled by pressing the 'backtick/tilde' key. This is done using
stb_image.h in a git submodule.

Unfortunately, stb_image.h does not conform to ANSI C, as it uses C++
style comments. To fix this and keep the project in strict C89, the
makefile includes a new target that strips comments from the file using
GNU sed. Some additional GCC warnings had to be suppressed in order to
include this file.