Render text as textured quads instead of bitmaps
Rendering text as textured quads increases performance by ~20-50% over
bitmap rasters. It also allows text to be scaled and moved by arbitrary
increments. As a consequence, text in 3D space needs to be reoriented
to face the player.

Another advantage is less jitter on the targeting reticules and being
able to correctly center their positions.