Added occlusion queries
This avoids drawing asteroids that aren't in view using a unit cube
as a bounding volume. The unit cube takes 192 bytes of buffer space.
Occlusion queries are disabled if the extension is not supported.

Added a 'const_data' value to A3DModel to account for static data that
does not need to be freed.